requirements.txt
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0